+++ Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+regex = "1.5"
 postgres = "0.19"
 serde = "1.0"
 serde_json = "1.0"
